More stuff.
* Multiple target frameworks.
* Fix a couple of memory issues.
* Documentation.
