+++ bindings/dotnet/ironcalc-dotnet/Directory.Build.props
@@ -1,15 +1,11 @@
 <Project DefaultTargets="Build">
  <PropertyGroup>
-   <TargetFramework>net9.0</TargetFramework>
+   <TargetFrameworks>net8.0;net9.0;net10.0</TargetFrameworks>
    <LangVersion>12</LangVersion>
    <Nullable>enable</Nullable>
-   <!--
    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
    <EnforceCodeStyleInBuild>true</EnforceCodeStyleInBuild>
    <EnableNETAnalyzers>true</EnableNETAnalyzers>
    <AnalysisModeStyle>Recommended</AnalysisModeStyle>
-   <GenerateDocumentationFile>true</GenerateDocumentationFile>
-   <NuGetAuditMode>direct</NuGetAuditMode>
-   -->
  </PropertyGroup>
 </Project>
