+++ bindings/dotnet/ironcalc-dotnet/IronCalc/Model.cs
@@ -1,52 +1,30 @@
 ﻿using System;
 using System.Runtime.CompilerServices;
 using System.Text;
+using System.Threading;
 using IronCalc.Native;
 
 namespace IronCalc;
 
-public class IronCalcException : Exception
-{
-    public readonly ErrorCode ErrorCode;
-
-    internal IronCalcException(string message, ErrorCode errorCode)
-        : base(message)
-    {
-        ErrorCode = errorCode;
-    }
-
-    internal IronCalcException(string message, ModelContextErrorTag? tag)
-        : base(message)
-    {
-        ErrorCode = tag switch {
-            null => ErrorCode.Unknown,
-            ModelContextErrorTag.XlsxError => ErrorCode.XslxError,
-            ModelContextErrorTag.WorkbookError => ErrorCode.WorkbookError,
-            ModelContextErrorTag.SetUserInputError => ErrorCode.SetUserInputError,
-            ModelContextErrorTag.GetUserInputError => ErrorCode.GetUserInputError,
-            _ => throw new ArgumentOutOfRangeException(nameof(tag), tag, null)
-        };
-    }
-}
-
-public enum ErrorCode
-{
-    Unknown = 1,
-    XslxError = 2,
-    WorkbookError = 3,
-    SetUserInputError = 4,
-    GetUserInputError = 5,
-}
-
+/// <summary>
+/// Represents an IronCalc spreadsheet model.
+/// </summary>
 public class Model : IDisposable
 {
-    private readonly unsafe ModelContext* ctx;
+    private readonly unsafe ModelContext* _ctx;
 
     private unsafe Model(ModelContext* ctx)
     {
-        this.ctx = ctx;
+        _ctx = ctx;
     }
 
+    /// <summary>
+    /// Creates a new empty spreadsheet model.
+    /// </summary>
+    /// <param name="name">The name of the workbook.</param>
+    /// <param name="locale">The locale to use for formula localization (e.g., "en_US").</param>
+    /// <param name="timezone">The IANA timezone to use for date/time functions (e.g., "UTC", "America/New_York").</param>
+    /// <returns>A new `Model` instance.</returns>
     public static Model NewEmpty(string name, string locale, string timezone)
     {
         unsafe
@@ -69,6 +47,14 @@ public class Model : IDisposable
         }
     }
 
+    /// <summary>
+    /// Loads a spreadsheet model from a byte array containing XLSX data.
+    /// </summary>
+    /// <param name="bytes">The byte array containing the XLSX file content.</param>
+    /// <param name="locale">The locale to use for formula localization (e.g., "en_US").</param>
+    /// <param name="timezone">The IANA timezone to use for date/time functions (e.g., "UTC", "America/New_York").</param>
+    /// <param name="name">An optional name for the workbook.</param>
+    /// <returns>A new `Model` instance.</returns>
     public static Model LoadFromXlsxBytes(byte[] bytes, string locale, string timezone, string? name = null)
     {
         unsafe
@@ -92,19 +78,29 @@ public class Model : IDisposable
         }
     }
 
+    /// <summary>
+    /// Evaluates all formulas in the spreadsheet.
+    /// </summary>
     public void Evaluate()
     {
         unsafe
         {
-            NativeMethods.evaluate(ctx);
+            NativeMethods.evaluate(_ctx);
         }
     }
 
+    /// <summary>
+    /// Gets the value of a cell by its sheet index, row, and column.
+    /// </summary>
+    /// <param name="sheet">The 0-based index of the sheet.</param>
+    /// <param name="row">The 1-based index of the row.</param>
+    /// <param name="column">The 1-based index of the column.</param>
+    /// <returns>A `CellValue` representing the value of the cell.</returns>
     public CellValue GetValue(uint sheet, int row, int column)
     {
         unsafe
         {
-            var result = NativeMethods.get_cell_value_by_index(ctx, sheet, row, column);
+            var result = NativeMethods.get_cell_value_by_index(_ctx, sheet, row, column);
             if (!result.is_ok)
             {
                 throw CreateExceptionFromError(result.error);
@@ -133,7 +129,7 @@ public class Model : IDisposable
                             Value = result.value->boolean_value,
                         };
                     default:
-                        throw new ArgumentOutOfRangeException();
+                        throw new InvalidOperationException($"Unknown tag: {result.value->tag}.");
                 }
             }
             finally
@@ -143,6 +139,13 @@ public class Model : IDisposable
         }
     }
 
+    /// <summary>
+    /// Sets the value of a cell.
+    /// </summary>
+    /// <param name="sheet">The 0-based index of the sheet.</param>
+    /// <param name="row">The 1-based index of the row.</param>
+    /// <param name="col">The 1-based index of the column.</param>
+    /// <param name="value">The value to set. If it starts with '=', it is treated as a formula.</param>
     public void SetUserInput(uint sheet, int row, int col, string value)
     {
         unsafe
@@ -150,7 +153,7 @@ public class Model : IDisposable
             var valueBytes = Encoding.UTF8.GetBytes(value);
             fixed (byte* valueP = valueBytes)
             {
-                var error = NativeMethods.set_user_input(ctx, sheet, row, col, valueP);
+                var error = NativeMethods.set_user_input(_ctx, sheet, row, col, valueP);
                 if (error != null)
                 {
                     throw CreateExceptionFromError(error);
@@ -163,26 +166,58 @@ public class Model : IDisposable
         ModelContextError* error,
         [CallerMemberName] string? callerName = null)
     {
-        string message;
+        var message = error->has_message
+            ? new string((sbyte*)error->message)
+            : $"Unknown error while calling {callerName ?? "UNKNOWN"} on IronCalc model.";
+
         var errorTag = error->tag;
-        if (error->has_message)
-        {
-            message = new String((sbyte*)error->message);
-            NativeMethods.dispose_error(error);
-        }
-        else
-        {
-            message = $"Unknown error while calling {callerName ?? "UNKNOWN"} on IronCalc model.";
-        }
+        NativeMethods.dispose_error(error);
 
         return new IronCalcException(message, errorTag);
     }
 
+    /// <summary>
+    /// The finalizer for <see cref="Model"/>.
+    /// This is called during garbage collection.
+    /// </summary>
+    ~Model()
+    {
+        Dispose(false);
+    }
+
+    /// <summary>
+    /// Releases the resources used by the `Model` instance.
+    /// </summary>
     public void Dispose()
     {
+        Dispose(true);
+        // Suppress finalization.
+        GC.SuppressFinalize(this);
+    }
+
+    private int _isDisposed;
+
+    /// <summary>
+    /// Releases the resources used by the `Model` instance.
+    /// </summary>
+    /// <param name="disposing"></param>
+    protected virtual void Dispose(bool disposing)
+    {
+        // In case _isDisposed is 0, atomically set it to 1.
+        // Enter the branch only if the original value is 1.
+        if (Interlocked.CompareExchange(ref _isDisposed, 1, 0) != 0)
+        {
+            return;
+        }
+
+        if (disposing)
+        {
+            // free managed resources.
+        }
+
         unsafe
         {
-            NativeMethods.dispose(ctx);
+            NativeMethods.dispose(_ctx);
         }
     }
 }
